requirements/base.txt
# Base requirements shared across all microservices
# This file is used by base/Dockerfile for building the base image
# Updated: 2024-01-XX - Consolidated from requirements-common.txt

# Core web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==2.3.3
flask-cors==4.0.0

# Data processing (core only)
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# HTTP client
requests>=2.32.4

# Configuration and validation
pydantic==2.5.0
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2

# Database (minimal) - Updated for Python 3.13 compatibility
sqlalchemy>=2.0.36

# Security and performance
cryptography>=42.0.5
psutil==5.9.6

# Production WSGI Server
gunicorn==21.2.0
